ci: fix failing lastfmstats test due to engine import
Since validate_lastfmstats is in the module from_lastfmstats,
which imports database engine indirectly, the pytest test fails in GitHub CI.
